refactor(coins): extract pure use cases and services to feature_coins_core
Slice 5a of #1111 — the mechanical half of the application layer. All 10
use cases plus the 6 services with no Flutter/platform dependencies move
to feature_coins_core; old paths keep re-export shims.

Deliberately left in app/ (real coupling, needs design decisions —
tracked for slice 5b):
- android_finance_import_preferences + coins_platform_support (Flutter
  and shared_preferences bound)
- cloud_mode_provider (AuthService/GoogleAuthService singletons)
- expense_providers (money feature's provider)
- the whole presentation layer (route_names, plus bill_split's parser
  service and itemized screen with their Gemini Nano / ML Kit deps)

Barrel hides the duplicate per-file `Result` typedefs the same way the
repository contracts already did; consolidating them into one shared
type is still open.

docs-not-needed: extraction behind shims, no behavior change.
