+++ app/lib/features/coins/application/services/coins_sync_service.dart
@@ -1,117 +1,1 @@
-/// Sync status for tracking offline operations
-enum SyncStatus { synced, pending, failed, conflict }
-
-/// Sync operation types
-enum SyncOperation { create, update, delete }
-
-/// Sync queue item
-class SyncQueueItem {
-  final String id;
-  final String entityType;
-  final String entityId;
-  final SyncOperation operation;
-  final Map<String, dynamic> data;
-  final DateTime createdAt;
-  final int retryCount;
-  final SyncStatus status;
-
-  const SyncQueueItem({
-    required this.id,
-    required this.entityType,
-    required this.entityId,
-    required this.operation,
-    required this.data,
-    required this.createdAt,
-    this.retryCount = 0,
-    this.status = SyncStatus.pending,
-  });
-}
-
-/// Sync service for Coins feature
-///
-/// Handles offline-first data synchronization using outbox pattern.
-///
-/// Phase: 1 & 2
-/// See: docs/features/coins/PROJECT_STRUCTURE.md
-abstract class CoinsSyncService {
-  /// Queue an operation for sync
-  Future<void> enqueue(SyncQueueItem item);
-
-  /// Process pending sync queue
-  Future<void> processQueue();
-
-  /// Get pending sync items count
-  Future<int> getPendingCount();
-
-  /// Get all pending items
-  Future<List<SyncQueueItem>> getPendingItems();
-
-  /// Clear completed/synced items
-  Future<void> clearCompleted();
-
-  /// Retry failed items
-  Future<void> retryFailed();
-
-  /// Check connectivity and trigger sync if online
-  Future<void> syncIfOnline();
-
-  /// Watch sync status changes
-  Stream<SyncStatus> watchSyncStatus();
-}
-
-/// Default implementation using local queue
-class CoinsSyncServiceImpl implements CoinsSyncService {
-  final List<SyncQueueItem> _queue = [];
-
-  @override
-  Future<void> enqueue(SyncQueueItem item) async {
-    // TODO: Persist to local database
-    _queue.add(item);
-  }
-
-  @override
-  Future<void> processQueue() async {
-    // TODO: Implement sync with backend
-    // 1. Check connectivity
-    // 2. Process items in order
-    // 3. Handle conflicts
-    // 4. Update status
-  }
-
-  @override
-  Future<int> getPendingCount() async {
-    return _queue.where((item) => item.status == SyncStatus.pending).length;
-  }
-
-  @override
-  Future<List<SyncQueueItem>> getPendingItems() async {
-    return _queue.where((item) => item.status == SyncStatus.pending).toList();
-  }
-
-  @override
-  Future<void> clearCompleted() async {
-    _queue.removeWhere((item) => item.status == SyncStatus.synced);
-  }
-
-  @override
-  Future<void> retryFailed() async {
-    // TODO: Reset failed items to pending and reprocess
-    for (var i = 0; i < _queue.length; i++) {
-      if (_queue[i].status == SyncStatus.failed) {
-        // Reset status to pending
-      }
-    }
-  }
-
-  @override
-  Future<void> syncIfOnline() async {
-    // TODO: Check connectivity using connectivity_plus package
-    // If online, call processQueue()
-  }
-
-  @override
-  Stream<SyncStatus> watchSyncStatus() async* {
-    // TODO: Implement status broadcast stream
-    yield SyncStatus.synced;
-  }
-}
+export 'package:feature_coins_core/src/application/services/coins_sync_service.dart';
