+++ app/test/features/coins/application/use_cases/add_group_member_use_case_test.dart
@@ -1,4 +1,4 @@
-import 'package:airo_app/features/coins/application/use_cases/add_group_member_use_case.dart';
+import 'package:feature_coins_core/src/application/use_cases/add_group_member_use_case.dart';
 import 'package:feature_coins_core/src/entities/group_member.dart';
 import 'package:airo_app/features/coins/domain/repositories/group_repository.dart'
     hide Result;
