+++ app/test/features/coins/presentation/screens/coins_dashboard_screen_test.dart
@@ -1,6 +1,6 @@
 import 'package:core_app_shell/core_app_shell.dart';
 import 'package:airo_app/features/coins/application/providers/android_import_permission_provider.dart';
-import 'package:airo_app/features/coins/application/services/android_finance_import_service.dart';
+import 'package:feature_coins_core/src/application/services/android_finance_import_service.dart';
 import 'package:airo_app/features/coins/application/providers/dashboard_providers.dart';
 import 'package:airo_app/features/coins/application/providers/expense_providers.dart';
 import 'package:feature_coins_core/src/entities/transaction.dart';
