+++ packages/feature_coins_core/lib/feature_coins_core.dart
@@ -1,45 +1,17 @@
-/// Pure domain layer of Airo Coins finance (money/budgets/groups):
-/// entities, value models, repository contracts, and calculation engines.
-/// Zero Flutter/app dependencies — see issue #1111 and ADR-0010.
-library;
-
-export 'src/entities/account.dart';
-export 'src/entities/budget.dart';
-export 'src/entities/category.dart';
-export 'src/entities/group.dart';
-export 'src/entities/group_member.dart';
-export 'src/entities/investment.dart';
-export 'src/entities/settlement.dart';
-export 'src/entities/shared_expense.dart';
-export 'src/entities/split_entry.dart';
-export 'src/entities/subscription.dart';
-export 'src/entities/transaction.dart';
-export 'src/errors/coins_errors.dart';
-export 'src/models/balance_summary.dart';
-export 'src/models/budget_status.dart';
-export 'src/models/currency.dart';
-export 'src/models/debt_entry.dart';
-export 'src/models/safe_to_spend.dart';
-export 'src/repositories/account_repository.dart';
-export 'src/repositories/budget_repository.dart' hide Result;
-export 'src/repositories/group_repository.dart' hide Result;
-export 'src/repositories/settlement_repository.dart' hide Result;
-export 'src/repositories/transaction_repository.dart' hide Result;
-export 'src/services/balance_engine.dart';
-export 'src/services/budget_engine.dart';
-export 'src/services/debt_simplifier.dart';
-export 'src/services/finance_insight_service.dart';
-export 'src/services/finance_message_parser.dart';
-export 'src/services/quick_add_expense_parser.dart';
-export 'src/services/split_calculator.dart';
-export 'src/data/datasources/coins_local_datasource.dart';
-export 'src/data/mappers/account_mapper.dart';
-export 'src/data/mappers/budget_mapper.dart';
-export 'src/data/mappers/group_mapper.dart';
-export 'src/data/mappers/settlement_mapper.dart';
-export 'src/data/mappers/transaction_mapper.dart';
-export 'src/data/repositories/account_repository_impl.dart';
-export 'src/data/repositories/budget_repository_impl.dart';
-export 'src/data/repositories/group_repository_impl.dart';
-export 'src/data/repositories/settlement_repository_impl.dart';
-export 'src/data/repositories/transaction_repository_impl.dart';
+export 'src/application/services/android_finance_import_service.dart';
+export 'src/application/services/coins_invite_link_service.dart';
+export 'src/application/services/coins_notification_service.dart';
+export 'src/application/services/coins_sync_service.dart';
+export 'src/application/services/finance_chat_ingestion_service.dart';
+export 'src/application/services/transaction_review_service.dart';
+export 'src/application/use_cases/add_expense_use_case.dart';
+export 'src/application/use_cases/add_group_member_use_case.dart' hide Result;
+export 'src/application/use_cases/add_split_use_case.dart' hide Result;
+export 'src/application/use_cases/calculate_balances_use_case.dart' hide Result;
+export 'src/application/use_cases/calculate_safe_to_spend_use_case.dart'
+    hide Result;
+export 'src/application/use_cases/create_group_use_case.dart' hide Result;
+export 'src/application/use_cases/delete_expense_use_case.dart' hide Result;
+export 'src/application/use_cases/record_settlement_use_case.dart' hide Result;
+export 'src/application/use_cases/set_budget_use_case.dart' hide Result;
+export 'src/application/use_cases/update_expense_use_case.dart' hide Result;
